% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{add_data}
\alias{add_data}
\title{Add data to propro}
\usage{
add_data(propro_model, name, value)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{name}{Name of data}

\item{value}{Value of data}
}
\description{
Add data to propro
}
