% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable.R
\name{add_variable}
\alias{add_variable}
\title{Add additional variable to propro}
\usage{
add_variable(propro_model, name, value)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{name}{Name of the variable}

\item{value}{Definition of the variable. Should start with <-}
}
\description{
Add additional variable to propro
}
