% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_consecutive_priors.R
\name{combine_consecutive_priors}
\alias{combine_consecutive_priors}
\title{Combine priors into one definition.}
\usage{
combine_consecutive_priors(propro_model, start, end)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{start}{First number of beta to combine}

\item{end}{Last beta to combine}
}
\description{
Useful for specificing certain priors, e.g. Dirichlet.
}
