% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bayesian_model.R
\name{create_propro}
\alias{create_propro}
\title{Create Probabilistic Model (propro) using event log and Petri net}
\usage{
create_propro(eventlog, PN)
}
\arguments{
\item{eventlog}{An eventlog object}

\item{PN}{A petri net object. Should have final marking.}
}
\description{
Create Probabilistic Model (propro) using event log and Petri net
}
