% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_priors.R
\name{list_priors}
\alias{list_priors}
\alias{list_traces}
\title{List priors and traces}
\usage{
list_priors(propro_model)

list_traces(propro_model)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}
}
\description{
List priors and traces
}
