% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{log_2_paper_ICPM}
\alias{log_2_paper_ICPM}
\title{Example log}
\format{Eventlog containing 50 artificial cases}
\usage{
log_2_paper_ICPM
}
\description{
Artifical eventlog
}
\keyword{datasets}
