% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{model_2_paper_ICPM}
\alias{model_2_paper_ICPM}
\title{Process model}
\format{Petri Net}
\usage{
model_2_paper_ICPM
}
\description{
Petri net describing log_2_paper_ICPM
}
\keyword{datasets}
