% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_automaton.R
\name{plot_automaton}
\alias{plot_automaton}
\title{Plot automaton}
\usage{
plot_automaton(propro_model)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}
}
\description{
Plot automaton
}
