% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propro.R
\docType{package}
\name{propro}
\alias{propro}
\alias{propro-package}
\title{propro - Creating probabilistic process models using Bayesian inference and MCMC}
\description{
Estimates posterior distributions for control-flow in a process model using Bayesian inference and McMC, starting from a Petri net and event log.
}
