% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_propro.R
\name{run_propro}
\alias{run_propro}
\title{Run propro model}
\usage{
run_propro(propro_model, n.chains, n.iter, n.burnin, ...)
}
\arguments{
\item{propro_model}{A propro model}

\item{n.chains}{Number of chains (see R2jags)}

\item{n.iter}{Number of iterations (see R2jargs)}

\item{n.burnin}{Burn in (see R2jags)}

\item{...}{Other arugements for R2jags}
}
\description{
Run propro model
}
