% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prior.R
\name{set_prior}
\alias{set_prior}
\title{Set prior}
\usage{
set_prior(propro_model, name, value)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{name}{Name of prior}

\item{value}{value of prior. Should start with <- or ~}
}
\description{
Set prior
}
