% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prior_complements.R
\name{set_prior_complements}
\alias{set_prior_complements}
\title{Set complements of prior}
\usage{
set_prior_complements(propro_model, n)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{n}{The amount of choices for which to set complements}
}
\description{
Set complements of prior
}
