% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_propro}
\alias{write_propro}
\title{Write model to file}
\usage{
write_propro(propro_model, file)
}
\arguments{
\item{propro_model}{A propro model, generated by create_propro}

\item{file}{File name}
}
\description{
Write model to file
}
